Put animations into their own script
Also might have broken dash cooldown but i'll fix that tomorrow in that case

+++ Assets/Scripts/Abilities/SCR_dash_cooldown.cs
@@ -14,7 +14,7 @@ public class SCR_dash_cooldown : MonoBehaviour
     Outline outline;
 
     private P_StateManager script;
-    private int dash_cooldown;
+    private float dash_cooldown;
 
     bool display_outline = false;
